+++ cli/src/commands/status.rs
@@ -50,7 +50,7 @@ use crate::ui::Ui;
 #[derive(clap::Args, Clone, Debug)]
 pub(crate) struct StatusArgs {
     /// Restrict the status display to these paths
-    #[arg(value_name = "FILESETS", value_hint = clap::ValueHint::AnyPath)]
+    #[arg(value_name = "FILESETS, value_hint = clap::ValueHint::AnyPath)]
     paths: Vec<String>,
 }