+++ cli/tests/test_bookmark_command.rs
@@ -2753,6 +2753,8 @@ fn test_bad_auto_track_bookmarks() {
     ]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
+    Auto-tracking 1 new file:
+      file
     Config error: Invalid `remotes.origin.auto-track-bookmarks`: Syntax error
     Caused by:  --> 1:1
       |