+++ cli/tests/test_operations.rs
@@ -1704,6 +1704,8 @@ fn test_op_diff_patch() {
     let output = work_dir.run_jj(["new"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
+    Auto-tracking 1 new file:
+      file
     Working copy  (@) now at: rlvkpnrz c1c924b8 (empty) (no description set)
     Parent commit (@-)      : qpvuntsm 6b57e33c (no description set)
     [EOF]
@@ -2771,6 +2773,8 @@ fn test_op_show_patch() {
     let output = work_dir.run_jj(["new"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
+    Auto-tracking 1 new file:
+      file
     Working copy  (@) now at: rlvkpnrz c1c924b8 (empty) (no description set)
     Parent commit (@-)      : qpvuntsm 6b57e33c (no description set)
     [EOF]